+++ exeqpdal/__init__.py
@@ -25,7 +25,7 @@ Usage:
 
 from __future__ import annotations
 
-__version__ = "0.1.0a5"
+__version__ = "0.1.0a6"
 __author__ = "ElGorrion"
 __license__ = "MIT"